MWPW-189894: unit test update, fetchResult update
